Добавила тест в RestApiPageTest.java для US_15.002 тест на проверку того что ссылка Rest Api открывается в той же вкладке удалила ненужный класс
